fix: align lead service identifier with MySQL bigint

+++ src/main/java/com/reno/leads/api/LeadResponse.java
@@ -1,3 +1,3 @@
 package com.reno.leads.api;
 
-public record LeadResponse(Long id, Integer service, String name, String email, String mobileNumber, String postalCode, String description, String status) {}
+public record LeadResponse(Long id, Long service, String name, String email, String mobileNumber, String postalCode, String description, String status) {}
